requirements.txt
pandas>=2.0.0
pdfplumber>=0.11.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pytest>=8.0.0
